+++ cmp/tab/TabModel.js
@@ -72,10 +72,6 @@ export class TabModel {
         this.refreshContextModel = new TabRefreshContextModel(this);
     }
 
-    isComponentModel() {
-        return true;
-    }
-
     activate() {
         this.containerModel.activateTab(this.id);
     }